refactor inspection read surface boundary
